feat: add Streamable HTTP MCP transport endpoint (/mcp)
Extract createMcpServer() factory into src/mcp-server.ts so both
stdio (src/index.ts) and HTTP (src/http.ts) share the same tool
definitions and handlers. The new /mcp route enables remote MCP
clients (e.g. OpenCode with type: remote) to connect over the
network, eliminating memory fragmentation between machines.

Stateless per-request — no session tracking needed since all
Engram tools are request/response.
